+++ csaf-matching/src/commonMain/kotlin/io/github/csaf/sbom/matching/Matcher.kt
@@ -16,8 +16,8 @@
  */
 package io.github.csaf.sbom.matching
 
-import io.github.csaf.sbom.matching.cpe.CPEMatchingTask
-import io.github.csaf.sbom.matching.purl.PurlMatchingTask
+import io.github.csaf.sbom.matching.old.CPEMatchingTask
+import io.github.csaf.sbom.matching.old.PurlMatchingTask
 import io.github.csaf.sbom.schema.generated.Csaf
 import io.github.csaf.sbom.validation.tests.affectedProducts
 import protobom.protobom.Document
