+++ csaf-matching/src/commonMain/kotlin/io/github/csaf/sbom/matching/Vers.kt
@@ -14,30 +14,10 @@
  * limitations under the License.
  *
  */
-package io.github.csaf.sbom.matching.purl
+package io.github.csaf.sbom.matching
 
-import protobom.protobom.Node
-import protobom.protobom.SoftwareIdentifierType
-
-expect class Purl(purl: String) {
-    fun canonicalize(): String
-
-    fun getScheme(): String?
-
-    fun getType(): String?
-
-    fun getNamespace(): String?
-
-    fun getName(): String?
-
-    fun getVersion(): String?
-
-    fun getQualifiers(): MutableMap<String, String>?
-
-    fun getSubpath(): String?
+expect class Vers {
+    fun contains(version: String): Boolean
 }
 
-val Node.purl: Purl?
-    get() {
-        return this.identifiers[SoftwareIdentifierType.PURL.value]?.let { Purl(it) }
-    }
+expect fun parseVers(versString: String): Vers?
