+++ csaf-matching/src/commonMain/kotlin/io/github/csaf/sbom/matching/properties/CpeProperty.kt
@@ -14,30 +14,12 @@
  * limitations under the License.
  *
  */
-package io.github.csaf.sbom.matching.cpe
+package io.github.csaf.sbom.matching.properties
 
+import io.github.csaf.sbom.matching.Cpe
 import io.github.csaf.sbom.matching.DefiniteMatch
 import io.github.csaf.sbom.matching.DefinitelyNoMatch
 import io.github.csaf.sbom.matching.MatchingConfidence
-import io.github.csaf.sbom.matching.properties.Property
-import io.github.csaf.sbom.matching.provider.PropertySource
-import protobom.protobom.Node
-import protobom.protobom.SoftwareIdentifierType
-
-expect interface Cpe {
-    fun matches(other: Cpe): Boolean
-
-    fun getVendor(): String
-}
-
-expect fun parseCpe(cpe: String): Cpe
-
-val Node.cpe: Cpe?
-    get() {
-        return (this.identifiers[SoftwareIdentifierType.CPE22.value]
-                ?: this.identifiers[SoftwareIdentifierType.CPE23.value])
-            ?.let { parseCpe(it) }
-    }
 
 /**
  * A property that represents a CPE value.
