+++ csaf-matching/src/commonMain/kotlin/io/github/csaf/sbom/matching/properties/VendorProperty.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2025, The Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package io.github.csaf.sbom.matching.properties
+
+import io.github.csaf.sbom.matching.Cpe
+import io.github.csaf.sbom.matching.Purl
+import io.github.csaf.sbom.matching.VulnerableProduct
+import io.github.csaf.sbom.schema.generated.Csaf
+import protobom.protobom.Node
+
+/** A property that presents a vendor. */
+typealias VendorProperty = StringProperty
+
+/**
+ * The [VendorPropertyProvider] is a [PropertyProvider] that provides the vendor of a product as a
+ * [VendorProperty].
+ *
+ * It extracts the vendor from the [VulnerableProduct.branches] or [Cpe] and returns it as a
+ * [VendorProperty].
+ */
+object VendorPropertyProvider : PropertyProvider<VendorProperty> {
+    override fun provideProperty(vulnerable: VulnerableProduct): VendorProperty? {
+        return vulnerable.branches
+            .firstOrNull { it.category == Csaf.Category3.vendor }
+            ?.name
+            ?.toProperty(PropertySource.OTHER)
+    }
+
+    override fun provideProperty(node: Node): VendorProperty? {
+        return null
+    }
+
+    override fun provideProperty(cpe: Cpe): VendorProperty? {
+        return cpe.getVendor().toProperty(PropertySource.CPE)
+    }
+
+    override fun provideProperty(purl: Purl): VendorProperty? {
+        return null
+    }
+}
